split announcement to its own component
